feat: exponential backoff and examples
