+++ retry_test.go
@@ -136,43 +136,3 @@ func (a *action) Call() error {
 	}
 	return err
 }
-
-func ExampleRetry_First() {
-	ctx, cancelFunc := context.WithTimeout(context.Background(), 10*time.Millisecond)
-	defer cancelFunc()
-
-	recoverErrorAction := &action{errors: []error{Recoverable(errors.New("failure"))}}
-
-	r := NewRetrier(
-		WithRetryPolicy(RetryRecoverablePolicy),
-		WithIntervalGenerator(StaticAttemptsGenerator(5, 1)))
-	r.Do(ctx, recoverErrorAction.Call)
-
-	// Output:
-	// action called 1 time(s)
-	// action called 2 time(s)
-	// action called 3 time(s)
-	// action called 4 time(s)
-	// action called 5 time(s)
-	// action called 6 time(s)
-}
-
-func ExampleRetry_Second() {
-	ctx, cancelFunc := context.WithTimeout(context.Background(), 10*time.Millisecond)
-	defer cancelFunc()
-
-	recoverErrorAction := &action{errors: []error{Recoverable(errors.New("failure"))}}
-
-	r := NewRetrier(
-		WithRetryPolicy(RetryRecoverablePolicy),
-		WithIntervalGenerator(BackoffGenerator))
-	r.Do(ctx, recoverErrorAction.Call)
-
-	// Output:
-	// action called 1 time(s)
-	// action called 2 time(s)
-	// action called 3 time(s)
-	// action called 4 time(s)
-	// action called 5 time(s)
-	// action called 6 time(s)
-}
